Initial Commit
Why are all these files coming in on the initial commit?

I had sensitive information in the previous git history, mainly related
to hardcoding different bits of information. I don't remember where all
it was, so it'd be very difficult to scrub it all effectively, so
instead I'm just re-making the git repository.